Add upstreams intelligently
